Added initial project structure
