+++ Behavioral.Automation.AsyncAbstractions.UI/BasicImplementations/ElementSelector.cs
@@ -1,4 +1,4 @@
-﻿namespace Behavioral.Automation.Bindings.UI.Abstractions;
+﻿namespace Behavioral.Automation.AsyncAbstractions.UI.BasicImplementations;
 
 public class ElementSelector
 {
