+++ Behavioral.Automation.AsyncAbstractions.UI/Interfaces/IPage.cs
@@ -1,0 +1,8 @@
+﻿namespace Behavioral.Automation.AsyncAbstractions.UI.Interfaces;
+
+public interface IPage
+{
+    public Task GoToUrlAsync(string url);
+    public Task GoToApplicationUrlAsync();
+    public Task HaveTitleAsync(string title);
+}
