+++ Behavioral.Automation.AsyncAbstractions.UI/Interfaces/IWebElement.cs
@@ -1,0 +1,11 @@
+using Behavioral.Automation.AsyncAbstractions.UI.BasicImplementations;
+
+namespace Behavioral.Automation.AsyncAbstractions.UI.Interfaces;
+
+public interface IWebElement
+{
+    public WebContext WebContext { get; }
+    public ElementSelector ElementSelector { get; }
+    public string? Description { get; set; }
+    public Task ShouldBecomeVisibleAsync();
+}
