+++ Behavioral.Automation.AsyncAbstractions.UI/Interfaces/IWebElementStorageService.cs
@@ -1,4 +1,4 @@
-﻿namespace Behavioral.Automation.Bindings.UI.Abstractions;
+﻿namespace Behavioral.Automation.AsyncAbstractions.UI.Interfaces;
 
 public interface IWebElementStorageService
 {
