--- Behavioral.Automation.Bindings.UI/Abstractions/IFileChooser.cs
@@ -1,6 +1,0 @@
-﻿namespace Behavioral.Automation.Bindings.UI.Abstractions;
-
-public interface IFileChooser
-{
-    public Task UploadFileAsync(string filePath);
-}
