--- Behavioral.Automation.Bindings.UI/Abstractions/IInputElement.cs
@@ -1,6 +1,0 @@
-﻿namespace Behavioral.Automation.Bindings.UI.Abstractions;
-
-public interface IInputElement : IWebElement
-{
-    public Task InputTextAsync(string text);
-}
