--- Behavioral.Automation.Bindings.UI/Bindings/InputBinding.cs
@@ -1,15 +1,0 @@
-using Behavioral.Automation.Bindings.UI.Abstractions;
-using TechTalk.SpecFlow;
-
-namespace Behavioral.Automation.Bindings.UI.Bindings;
-
-[Binding]
-public class InputBinding
-{
-    [Given(@"""(.+?)"" text was entered into ""(.+?)"" input")]
-    [When(@"user enters ""(.+?)"" into ""(.+?)"" input")]
-    public async Task FillInput(string text, IInputElement input)
-    {
-        await input.InputTextAsync(text);
-    }
-}
