--- Behavioral.Automation.Bindings.UI/CHANGELOG.md
@@ -1,11 +1,0 @@
-# Changelog
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-[0.0.1] - 17-02-2024
-### Added
-- A few Button, Checkbox, Input, Label, Navigation, Page and Table bindings
-- Abstractions for WebElements and Configuration classes (for example, IWebElementStorageService)
-- ElementTransformations class
